+++ src/main/java/com/midnear/midnearshopping/mapper/storeManagement/StatisticsMapper.java
@@ -3,14 +3,15 @@ package com.midnear.midnearshopping.mapper.storeManagement;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Mapper
 public interface StatisticsMapper {
-    Long getDailySales(Date date);
-    Long getRefundedTotalPrice(Date date);
-    Long getWeeklySales(@Param("start") String start, @Param("end") String end);
-    Long getWeeklyRefundedTotalPrice(@Param("start") String start, @Param("end") String end);
-    Long getMonthlySales(String month);
-    Long getMonthlyRefundedTotalPrice(String month);
+    BigDecimal getDailySales(Date date);
+    BigDecimal getRefundedTotalPrice(Date date);
+    BigDecimal getWeeklySales(@Param("start") String start, @Param("end") String end);
+    BigDecimal getWeeklyRefundedTotalPrice(@Param("start") String start, @Param("end") String end);
+    BigDecimal getMonthlySales(String month);
+    BigDecimal getMonthlyRefundedTotalPrice(String month);
 }
